requirements.txt
Django>=4.2.0,<4.3
djangorestframework>=3.14.0,<3.15
requests>=2.31.0
django-cors-headers>=4.3.1
google-genai